(#11060) Minor fix to a lab
Hi,

We have this workshop *manage-pdb* in Oracle Database 21c, `WMS ID #11060`. The workshop is published to production recently.
Now, while testing the green button environment, we came across a minor issue in one of the steps and updated the lab - `provision a PDB`.
Please approve the PR so we can enable the green button. Thank you
~MG